pyproject.toml
[project]
name = "vitaprimo"
version = "0.1.0"
description = "game in pygame-ce"
requires-python = ">=3.13"
dependencies = ["pygame-ce"]
